+++ changelog.md
@@ -4,6 +4,10 @@ This project does NOT (yet) adhere to [Semantic Versioning](http://semver.org/).
 It is running as an Alfa/Beta, once v1 is reached we will start supporting semver.
 For now, please use all minor versions as major versions and each build version will be a minor version.
 
+## [Unreleased]
+## Added
+- `deliveryCost` to create order call
+
 ## [0.5.2]
 ## Added
 - `JsonSerializable` to all entities - all entities can now be serialized to JSON